fix: Move auth layout to auth layout component
Move the layout to conditionally render children inside the Auth
component;
