+++ src/libs/components/app/app.tsx
@@ -1,7 +1,7 @@
 import { Auth } from "~/pages/auth/auth.tsx";
 
 function App() {
-  return <Auth> </Auth>;
+  return <Auth />;
 }
 
 export { App };
