+++ src/pages/auth/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { AuthLayout } from "./auth-layout/auth-layout.tsx";
